+++ app/Http/Controllers/EquipmentFileController.php
@@ -2,12 +2,14 @@
 
 namespace App\Http\Controllers;
 
+use App\User;
 use App\Equipment;
-use App\Enums\Permissions;
+use App\Enums\Perm;
 use Illuminate\Http\Request;
 use App\Http\Helpers\FileHelper;
 use App\Http\Helpers\FilesHelper;
 use App\Http\Requests\FileRequest;
+use Illuminate\Support\Facades\Gate;
 use App\Events\EquipmentFiles\ECreate;
 use App\Events\EquipmentFiles\EDelete;
 use App\Events\EquipmentFiles\EUpdate;
@@ -16,6 +18,16 @@ use Illuminate\Support\Facades\Storage;
 class EquipmentFileController extends Controller
 {
     /**
+     * @var Equipment
+     */
+    private $_equipment;
+
+    /**
+     * @var User
+     */
+    private $_user;
+
+    /**
      * Add middleware depends on user permissions.
      *
      * @param  Request  $request
@@ -23,12 +35,30 @@ class EquipmentFileController extends Controller
      */
     public function permissions(Request $request): array
     {
+        $this->_user = auth()->user();
+
+        if (! $this->_user) {
+            $this->middleware('jwt.auth');
+            return [];
+        }
+
+        $equipmentId = (int) $request->route('equipment');
+        $this->_equipment = Equipment::findOrFail($equipmentId);
+
+        // Permissions on equipment before get a files
+        if (! $this->_user->perm(Perm::EQUIPMENTS_VIEW_ALL) &&
+            Gate::denies('owner', $this->_equipment)
+        ) {
+            $this->middleware('permission:disable');
+            return [];
+        }
+
         return [
-            'index' => Permissions::EQUIPMENTS_FILES_VIEW,
-            'show' => Permissions::EQUIPMENTS_FILES_DOWNLOAD,
-            'store' => Permissions::EQUIPMENTS_FILES_CREATE,
-            'update' => Permissions::EQUIPMENTS_FILES_EDIT,
-            'destroy' => Permissions::EQUIPMENTS_FILES_DELETE,
+            'index' => [Perm::EQUIPMENTS_FILES_VIEW_OWN, Perm::EQUIPMENTS_FILES_VIEW_ALL],
+            'show' => [Perm::EQUIPMENTS_FILES_VIEW_OWN, Perm::EQUIPMENTS_FILES_VIEW_ALL],
+            'store' => Perm::EQUIPMENTS_FILES_CREATE,
+            'update' => [Perm::EQUIPMENTS_FILES_EDIT_OWN, Perm::EQUIPMENTS_FILES_EDIT_ALL],
+            'destroy' => [Perm::EQUIPMENTS_FILES_DELETE_OWN, Perm::EQUIPMENTS_FILES_DELETE_ALL],
         ];
     }
 
@@ -40,11 +70,15 @@ class EquipmentFileController extends Controller
      */
     public function index(int $equipmentId)
     {
-        $equipment = Equipment::findOrFail($equipmentId);
+        $equipmentFiles = $this->_equipment->files();
+
+        if (! $this->_user->perm(Perm::EQUIPMENTS_FILES_VIEW_ALL)) {
+            $equipmentFiles->where('user_id', $this->_user->id);
+        }
 
         return response()->json([
             'message' => __('app.files.files_get'),
-            'equipment_files' => $equipment->files,
+            'equipment_files' => $equipmentFiles->get(),
         ]);
     }
 
@@ -57,18 +91,17 @@ class EquipmentFileController extends Controller
      */
     public function store(FileRequest $request, int $equipmentId)
     {
-        $equipment = Equipment::findOrFail($equipmentId);
         $requestFiles = $request->file('files');
 
         $filesHelper = new FilesHelper($requestFiles);
         $filesHelper->upload('equipments/'.$equipmentId);
 
         $uploadedIds = $filesHelper->getUploadedIds();
-        $equipment->files()->attach($uploadedIds);
-        $uploadedFiles = $equipment->files()->whereIn('files.id', $uploadedIds)->get();
+        $this->_equipment->files()->attach($uploadedIds);
+        $uploadedFiles = $this->_equipment->files()->whereIn('files.id', $uploadedIds)->get();
 
         if (count($uploadedFiles)) {
-            event(new ECreate($equipmentId, $uploadedFiles->toArray()));
+            event(new ECreate($equipmentId, $uploadedFiles));
         }
 
         if ($filesHelper->hasErrors()) {
@@ -94,9 +127,14 @@ class EquipmentFileController extends Controller
      */
     public function show(int $equipmentId, int $fileId)
     {
-        $equipmentFile = Equipment::findOrFail($equipmentId)
-            ->files()
-            ->findOrFail($fileId);
+        $equipmentFile = $this->_equipment->files()->findOrFail($fileId);
+
+        // Show only own file
+        if (! $this->_user->perm(Perm::EQUIPMENTS_FILES_VIEW_ALL) &&
+            Gate::denies('owner', $equipmentFile)
+        ) {
+            return $this->responseNoPermission();
+        }
 
         if (! Storage::exists($equipmentFile->file)) {
             return response()->json(['message' => __('app.files.file_not_found')], 422);
@@ -115,14 +153,19 @@ class EquipmentFileController extends Controller
      */
     public function update(FileRequest $request, int $equipmentId, int $fileId)
     {
-        $equipmentFile = Equipment::findOrFail($equipmentId)
-            ->files()
-            ->findOrFail($fileId);
+        $equipmentFile = $this->_equipment->files()->findOrFail($fileId);
+
+        // Edit only own file
+        if (! $this->_user->perm(Perm::EQUIPMENTS_FILES_EDIT_ALL) &&
+            Gate::denies('owner', $equipmentFile)
+        ) {
+            return $this->responseNoPermission();
+        }
 
         $equipmentFile->name = $request->name;
 
         if (! $equipmentFile->save()) {
-            return response()->json(['message' => __('app.database.save_error')], 422);
+            return $this->responseDatabaseSaveError();
         }
 
         event(new EUpdate($equipmentId, $fileId, $equipmentFile));
@@ -139,17 +182,23 @@ class EquipmentFileController extends Controller
      * @param  int  $equipmentId
      * @param  int  $fileId
      * @return \Illuminate\Http\JsonResponse
+     * @throws \Exception
      */
     public function destroy(int $equipmentId, int $fileId)
     {
-        $equipmentFile = Equipment::findOrFail($equipmentId)
-            ->files()
-            ->findOrFail($fileId);
+        $equipmentFile = $this->_equipment->files()->findOrFail($fileId);
+
+        // Delete only own file
+        if (! $this->_user->perm(Perm::EQUIPMENTS_FILES_DELETE_ALL) &&
+            Gate::denies('owner', $equipmentFile)
+        ) {
+            return $this->responseNoPermission();
+        }
 
         FileHelper::delete($equipmentFile->file);
 
         if (! $equipmentFile->delete()) {
-            return response()->json(['message' => __('app.database.destroy_error')], 422);
+            return $this->responseDatabaseDestroyError();
         }
 
         event(new EDelete($equipmentId, $fileId));
